% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{wlbb}
\alias{wlbb}
\title{WLBB_plot_stock}
\usage{
wlbb(symbol)
}
\arguments{
\item{symbol}{The symbol of stock, "600000.ss" for example.}
}
\value{
a plot
}
\description{
imput a symbol of a stock,return a plot.
}
\author{
wuliubingbing@foxmail.com
}
